refactor: use referrer interface, gateway diff and gateway wrapper
