+++ controllers/limitador_cluster_envoyfilter_controller.go
@@ -22,6 +22,7 @@ import (
 	"fmt"
 
 	"github.com/go-logr/logr"
+	"github.com/kuadrant/kuadrant-operator/api/v1beta2"
 	limitadorv1alpha1 "github.com/kuadrant/limitador-operator/api/v1alpha1"
 	istioapinetworkingv1alpha3 "istio.io/api/networking/v1alpha3"
 	istioclientnetworkingv1alpha3 "istio.io/client-go/pkg/apis/networking/v1alpha3"
@@ -119,8 +120,8 @@ func (r *LimitadorClusterEnvoyFilterReconciler) desiredRateLimitingClusterEnvoyF
 		},
 	}
 
-	gateway := common.GatewayWrapper{Gateway: gw, PolicyRefsConfig: &common.KuadrantRateLimitPolicyRefsConfig{}}
-	rlpRefs := gateway.PolicyRefs()
+	gateway := reconcilers.GatewayWrapper{Gateway: gw, Referrer: &v1beta2.RateLimitPolicy{}}
+	rlpRefs := common.BackReferencesFromObject(gateway.Gateway, gateway.Referrer)
 	logger.V(1).Info("desiredRateLimitingClusterEnvoyFilter", "rlpRefs", rlpRefs)
 
 	if len(rlpRefs) < 1 {
