+++ pkg/common/gatewayapi_utils.go
@@ -2,12 +2,10 @@ package common
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"reflect"
 	"strings"
 
-	"golang.org/x/exp/slices"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/types"
@@ -294,227 +292,6 @@ func routePathMatchToRulePath(pathMatch *gatewayapiv1beta1.HTTPPathMatch) []stri
 	}
 
 	return []string{val + suffix}
-}
-
-type PolicyRefsConfig interface {
-	PolicyRefsAnnotation() string
-}
-
-type KuadrantRateLimitPolicyRefsConfig struct{}
-
-func (c *KuadrantRateLimitPolicyRefsConfig) PolicyRefsAnnotation() string {
-	return RateLimitPoliciesBackRefAnnotation
-}
-
-type KuadrantAuthPolicyRefsConfig struct{}
-
-func (c *KuadrantAuthPolicyRefsConfig) PolicyRefsAnnotation() string {
-	return AuthPoliciesBackRefAnnotation
-}
-
-func GatewaysMissingPolicyRef(gwList *gatewayapiv1beta1.GatewayList, policyKey client.ObjectKey, policyGwKeys []client.ObjectKey, config PolicyRefsConfig) []GatewayWrapper {
-	// gateways referenced by the policy but do not have reference to it in the annotations
-	gateways := make([]GatewayWrapper, 0)
-	for i := range gwList.Items {
-		gateway := gwList.Items[i]
-		gw := GatewayWrapper{&gateway, config}
-		if slices.Contains(policyGwKeys, client.ObjectKeyFromObject(&gateway)) && !gw.ContainsPolicy(policyKey) {
-			gateways = append(gateways, gw)
-		}
-	}
-	return gateways
-}
-
-func GatewaysWithValidPolicyRef(gwList *gatewayapiv1beta1.GatewayList, policyKey client.ObjectKey, policyGwKeys []client.ObjectKey, config PolicyRefsConfig) []GatewayWrapper {
-	// gateways referenced by the policy but also have reference to it in the annotations
-	gateways := make([]GatewayWrapper, 0)
-	for i := range gwList.Items {
-		gateway := gwList.Items[i]
-		gw := GatewayWrapper{&gateway, config}
-		if slices.Contains(policyGwKeys, client.ObjectKeyFromObject(&gateway)) && gw.ContainsPolicy(policyKey) {
-			gateways = append(gateways, gw)
-		}
-	}
-	return gateways
-}
-
-func GatewaysWithInvalidPolicyRef(gwList *gatewayapiv1beta1.GatewayList, policyKey client.ObjectKey, policyGwKeys []client.ObjectKey, config PolicyRefsConfig) []GatewayWrapper {
-	// gateways not referenced by the policy but still have reference in the annotations
-	gateways := make([]GatewayWrapper, 0)
-	for i := range gwList.Items {
-		gateway := gwList.Items[i]
-		gw := GatewayWrapper{&gateway, config}
-		if !slices.Contains(policyGwKeys, client.ObjectKeyFromObject(&gateway)) && gw.ContainsPolicy(policyKey) {
-			gateways = append(gateways, gw)
-		}
-	}
-	return gateways
-}
-
-// GatewayWrapper wraps a Gateway API Gateway adding methods and configs to manage policy references in annotations
-type GatewayWrapper struct {
-	*gatewayapiv1beta1.Gateway
-	PolicyRefsConfig
-}
-
-func (g GatewayWrapper) Key() client.ObjectKey {
-	if g.Gateway == nil {
-		return client.ObjectKey{}
-	}
-	return client.ObjectKeyFromObject(g.Gateway)
-}
-
-func (g GatewayWrapper) PolicyRefs() []client.ObjectKey {
-	if g.Gateway == nil {
-		return make([]client.ObjectKey, 0)
-	}
-
-	gwAnnotations := ReadAnnotationsFromObject(g)
-
-	val, ok := gwAnnotations[g.PolicyRefsAnnotation()]
-	if !ok {
-		return make([]client.ObjectKey, 0)
-	}
-
-	var refs []client.ObjectKey
-
-	err := json.Unmarshal([]byte(val), &refs)
-	if err != nil {
-		return make([]client.ObjectKey, 0)
-	}
-
-	return refs
-}
-
-func (g GatewayWrapper) ContainsPolicy(policyKey client.ObjectKey) bool {
-	if g.Gateway == nil {
-		return false
-	}
-
-	gwAnnotations := ReadAnnotationsFromObject(g)
-
-	val, ok := gwAnnotations[g.PolicyRefsAnnotation()]
-	if !ok {
-		return false
-	}
-
-	var refs []client.ObjectKey
-
-	err := json.Unmarshal([]byte(val), &refs)
-	if err != nil {
-		return false
-	}
-
-	return slices.Contains(refs, policyKey)
-}
-
-// AddPolicy tries to add a policy to the existing ref list.
-// Returns true if policy was added, false otherwise
-func (g GatewayWrapper) AddPolicy(policyKey client.ObjectKey) bool {
-	if g.Gateway == nil {
-		return false
-	}
-
-	gwAnnotations := ReadAnnotationsFromObject(g)
-
-	val, ok := gwAnnotations[g.PolicyRefsAnnotation()]
-	if !ok {
-		refs := []client.ObjectKey{policyKey}
-		serialized, err := json.Marshal(refs)
-		if err != nil {
-			return false
-		}
-		gwAnnotations[g.PolicyRefsAnnotation()] = string(serialized)
-		g.SetAnnotations(gwAnnotations)
-		return true
-	}
-
-	var refs []client.ObjectKey
-
-	err := json.Unmarshal([]byte(val), &refs)
-	if err != nil {
-		return false
-	}
-
-	if slices.Contains(refs, policyKey) {
-		return false
-	}
-
-	refs = append(refs, policyKey)
-	serialized, err := json.Marshal(refs)
-	if err != nil {
-		return false
-	}
-	gwAnnotations[g.PolicyRefsAnnotation()] = string(serialized)
-	g.SetAnnotations(gwAnnotations)
-	return true
-}
-
-// DeletePolicy tries to delete a policy from the existing ref list.
-// Returns true if the policy was deleted, false otherwise
-func (g GatewayWrapper) DeletePolicy(policyKey client.ObjectKey) bool {
-	if g.Gateway == nil {
-		return false
-	}
-
-	gwAnnotations := ReadAnnotationsFromObject(g)
-
-	val, ok := gwAnnotations[g.PolicyRefsAnnotation()]
-	if !ok {
-		return false
-	}
-
-	var refs []client.ObjectKey
-
-	err := json.Unmarshal([]byte(val), &refs)
-	if err != nil {
-		return false
-	}
-
-	if refID := FindObjectKey(refs, policyKey); refID != len(refs) {
-		// remove index
-		refs = append(refs[:refID], refs[refID+1:]...)
-		serialized, err := json.Marshal(refs)
-		if err != nil {
-			return false
-		}
-		gwAnnotations[g.PolicyRefsAnnotation()] = string(serialized)
-		g.SetAnnotations(gwAnnotations)
-		return true
-	}
-
-	return false
-}
-
-// Hostnames builds a list of hostnames from the listeners.
-func (g GatewayWrapper) Hostnames() []gatewayapiv1beta1.Hostname {
-	hostnames := make([]gatewayapiv1beta1.Hostname, 0)
-	if g.Gateway == nil {
-		return hostnames
-	}
-
-	for idx := range g.Spec.Listeners {
-		if g.Spec.Listeners[idx].Hostname != nil {
-			hostnames = append(hostnames, *g.Spec.Listeners[idx].Hostname)
-		}
-	}
-
-	return hostnames
-}
-
-// GatewayWrapperList is a list of GatewayWrappers that implements sort.Interface
-type GatewayWrapperList []GatewayWrapper
-
-func (g GatewayWrapperList) Len() int {
-	return len(g)
-}
-
-func (g GatewayWrapperList) Less(i, j int) bool {
-	return g[i].CreationTimestamp.Before(&g[j].CreationTimestamp)
-}
-
-func (g GatewayWrapperList) Swap(i, j int) {
-	g[i], g[j] = g[j], g[i]
 }
 
 // TargetHostnames returns an array of hostnames coming from the network object (HTTPRoute, Gateway)
