+++ pkg/common/gatewayapi_utils_test.go
@@ -8,11 +8,9 @@ import (
 	"reflect"
 	"testing"
 
-	"golang.org/x/exp/slices"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	k8stypes "k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	gatewayapiv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
@@ -679,384 +677,6 @@ func TestHTTPRouteRuleToString(t *testing.T) {
 	}
 }
 
-func TestGatewaysMissingPolicyRef(t *testing.T) {
-	gwList := &gatewayapiv1beta1.GatewayList{
-		Items: []gatewayapiv1beta1.Gateway{
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace:   "gw-ns",
-					Name:        "gw-1",
-					Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-				},
-			},
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace:   "gw-ns",
-					Name:        "gw-2",
-					Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"}]`},
-				},
-			},
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "gw-ns",
-					Name:      "gw-3",
-				},
-			},
-		},
-	}
-
-	var gws []string
-	policyRefConfig := &KuadrantRateLimitPolicyRefsConfig{}
-	gwName := func(gw GatewayWrapper) string { return gw.Gateway.Name }
-
-	gws = Map(GatewaysMissingPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-1"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-2"},
-		{Namespace: "gw-ns", Name: "gw-3"},
-	}, policyRefConfig), gwName)
-
-	if slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected not to be listed as missing policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as missing policy ref")
-	}
-	if !slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected to be listed as missing policy ref")
-	}
-
-	gws = Map(GatewaysMissingPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-2"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-1"},
-	}, policyRefConfig), gwName)
-
-	if slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected not to be listed as missing policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as missing policy ref")
-	}
-	if slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected not to be listed as missing policy ref")
-	}
-
-	gws = Map(GatewaysMissingPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-3"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-1"},
-		{Namespace: "gw-ns", Name: "gw-3"},
-	}, policyRefConfig), gwName)
-
-	if !slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected to be listed as missing policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as missing policy ref")
-	}
-	if !slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected to be listed as missing policy ref")
-	}
-}
-
-func TestGatewaysWithValidPolicyRef(t *testing.T) {
-	gwList := &gatewayapiv1beta1.GatewayList{
-		Items: []gatewayapiv1beta1.Gateway{
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace:   "gw-ns",
-					Name:        "gw-1",
-					Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-				},
-			},
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace:   "gw-ns",
-					Name:        "gw-2",
-					Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"}]`},
-				},
-			},
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "gw-ns",
-					Name:      "gw-3",
-				},
-			},
-		},
-	}
-
-	var gws []string
-	policyRefConfig := &KuadrantRateLimitPolicyRefsConfig{}
-	gwName := func(gw GatewayWrapper) string { return gw.Gateway.Name }
-
-	gws = Map(GatewaysWithValidPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-1"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-2"},
-		{Namespace: "gw-ns", Name: "gw-3"},
-	}, policyRefConfig), gwName)
-
-	if slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected not to be listed as with valid policy ref")
-	}
-	if !slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected to be listed as with valid policy ref")
-	}
-	if slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected not to be listed as with valid policy ref")
-	}
-
-	gws = Map(GatewaysWithValidPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-2"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-1"},
-	}, policyRefConfig), gwName)
-
-	if !slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected to be listed as with valid policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as with valid policy ref")
-	}
-	if slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected not to be listed as with valid policy ref")
-	}
-
-	gws = Map(GatewaysWithValidPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-3"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-1"},
-		{Namespace: "gw-ns", Name: "gw-3"},
-	}, policyRefConfig), gwName)
-
-	if slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected not to be listed as with valid policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as with valid policy ref")
-	}
-	if slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected not to be listed as with valid policy ref")
-	}
-}
-
-func TestGatewaysWithInvalidPolicyRef(t *testing.T) {
-	gwList := &gatewayapiv1beta1.GatewayList{
-		Items: []gatewayapiv1beta1.Gateway{
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace:   "gw-ns",
-					Name:        "gw-1",
-					Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-				},
-			},
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace:   "gw-ns",
-					Name:        "gw-2",
-					Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"}]`},
-				},
-			},
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "gw-ns",
-					Name:      "gw-3",
-				},
-			},
-		},
-	}
-
-	var gws []string
-	policyRefConfig := &KuadrantRateLimitPolicyRefsConfig{}
-	gwName := func(gw GatewayWrapper) string { return gw.Gateway.Name }
-
-	gws = Map(GatewaysWithInvalidPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-1"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-2"},
-		{Namespace: "gw-ns", Name: "gw-3"},
-	}, policyRefConfig), gwName)
-
-	if !slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected to be listed as with invalid policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-	if slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-
-	gws = Map(GatewaysWithInvalidPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-2"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-1"},
-	}, policyRefConfig), gwName)
-
-	if slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-	if slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-
-	gws = Map(GatewaysWithInvalidPolicyRef(gwList, k8stypes.NamespacedName{Namespace: "app-ns", Name: "policy-3"}, []client.ObjectKey{
-		{Namespace: "gw-ns", Name: "gw-1"},
-		{Namespace: "gw-ns", Name: "gw-3"},
-	}, policyRefConfig), gwName)
-
-	if slices.Contains(gws, "gw-1") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-	if slices.Contains(gws, "gw-2") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-	if slices.Contains(gws, "gw-3") {
-		t.Error("gateway expected not to be listed as with invalid policy ref")
-	}
-}
-
-func TestGatewayWrapperKey(t *testing.T) {
-	gw := GatewayWrapper{
-		Gateway: &gatewayapiv1beta1.Gateway{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace:   "gw-ns",
-				Name:        "gw-1",
-				Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-			},
-		},
-		PolicyRefsConfig: &KuadrantRateLimitPolicyRefsConfig{},
-	}
-	if gw.Key().Namespace != "gw-ns" || gw.Key().Name != "gw-1" {
-		t.Fail()
-	}
-}
-
-func TestGatewayWrapperPolicyRefs(t *testing.T) {
-	gw := GatewayWrapper{
-		Gateway: &gatewayapiv1beta1.Gateway{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace:   "gw-ns",
-				Name:        "gw-1",
-				Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-			},
-		},
-		PolicyRefsConfig: &KuadrantRateLimitPolicyRefsConfig{},
-	}
-	refs := Map(gw.PolicyRefs(), func(ref k8stypes.NamespacedName) string { return ref.String() })
-	if !slices.Contains(refs, "app-ns/policy-1") {
-		t.Error("GatewayWrapper.PolicyRefs() should contain app-ns/policy-1")
-	}
-	if !slices.Contains(refs, "app-ns/policy-2") {
-		t.Error("GatewayWrapper.PolicyRefs() should contain app-ns/policy-2")
-	}
-	if len(refs) != 2 {
-		t.Fail()
-	}
-}
-
-func TestGatewayWrapperContainsPolicy(t *testing.T) {
-	gw := GatewayWrapper{
-		Gateway: &gatewayapiv1beta1.Gateway{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace:   "gw-ns",
-				Name:        "gw-1",
-				Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-			},
-		},
-		PolicyRefsConfig: &KuadrantRateLimitPolicyRefsConfig{},
-	}
-	if !gw.ContainsPolicy(client.ObjectKey{Namespace: "app-ns", Name: "policy-1"}) {
-		t.Error("GatewayWrapper.ContainsPolicy() should contain app-ns/policy-1")
-	}
-	if !gw.ContainsPolicy(client.ObjectKey{Namespace: "app-ns", Name: "policy-2"}) {
-		t.Error("GatewayWrapper.ContainsPolicy() should contain app-ns/policy-1")
-	}
-	if gw.ContainsPolicy(client.ObjectKey{Namespace: "app-ns", Name: "policy-3"}) {
-		t.Error("GatewayWrapper.ContainsPolicy() should not contain app-ns/policy-1")
-	}
-}
-
-func TestGatewayWrapperAddPolicy(t *testing.T) {
-	gateway := gatewayapiv1beta1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace:   "gw-ns",
-			Name:        "gw-1",
-			Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-		},
-	}
-	gw := GatewayWrapper{
-		Gateway:          &gateway,
-		PolicyRefsConfig: &KuadrantRateLimitPolicyRefsConfig{},
-	}
-	if gw.AddPolicy(client.ObjectKey{Namespace: "app-ns", Name: "policy-1"}) {
-		t.Error("GatewayWrapper.AddPolicy() expected to return false")
-	}
-	if !gw.AddPolicy(client.ObjectKey{Namespace: "app-ns", Name: "policy-3"}) {
-		t.Error("GatewayWrapper.AddPolicy() expected to return true")
-	}
-	if gw.Annotations["kuadrant.io/ratelimitpolicies"] != `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"},{"Namespace":"app-ns","Name":"policy-3"}]` {
-		t.Error("GatewayWrapper.AddPolicy() expected to have added policy ref to the annotations")
-	}
-}
-
-func TestGatewayDeletePolicy(t *testing.T) {
-	gateway := gatewayapiv1beta1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace:   "gw-ns",
-			Name:        "gw-1",
-			Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-		},
-	}
-	gw := GatewayWrapper{
-		Gateway:          &gateway,
-		PolicyRefsConfig: &KuadrantRateLimitPolicyRefsConfig{},
-	}
-	if !gw.DeletePolicy(client.ObjectKey{Namespace: "app-ns", Name: "policy-1"}) {
-		t.Error("GatewayWrapper.DeletePolicy() expected to return true")
-	}
-	if gw.DeletePolicy(client.ObjectKey{Namespace: "app-ns", Name: "policy-3"}) {
-		t.Error("GatewayWrapper.DeletePolicy() expected to return false")
-	}
-	if gw.Annotations["kuadrant.io/ratelimitpolicies"] != `[{"Namespace":"app-ns","Name":"policy-2"}]` {
-		t.Error("GatewayWrapper.DeletePolicy() expected to have deleted policy ref from the annotations")
-	}
-}
-
-func TestGatewayHostnames(t *testing.T) {
-	hostname := gatewayapiv1beta1.Hostname("toystore.com")
-	gateway := gatewayapiv1beta1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace:   "gw-ns",
-			Name:        "gw-1",
-			Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-		},
-		Spec: gatewayapiv1beta1.GatewaySpec{
-			Listeners: []gatewayapiv1beta1.Listener{
-				{
-					Name:     "my-listener",
-					Hostname: &hostname,
-				},
-			},
-		},
-	}
-	gw := GatewayWrapper{
-		Gateway:          &gateway,
-		PolicyRefsConfig: &KuadrantRateLimitPolicyRefsConfig{},
-	}
-	hostnames := gw.Hostnames()
-	if !slices.Contains(hostnames, "toystore.com") {
-		t.Error("GatewayWrapper.Hostnames() expected to contain 'toystore.com'")
-	}
-	if len(hostnames) != 1 {
-		t.Fail()
-	}
-}
-
-func TestGatewayWrapperPolicyRefsAnnotation(t *testing.T) {
-	gw := GatewayWrapper{
-		Gateway: &gatewayapiv1beta1.Gateway{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace:   "gw-ns",
-				Name:        "gw-1",
-				Annotations: map[string]string{"kuadrant.io/ratelimitpolicies": `[{"Namespace":"app-ns","Name":"policy-1"},{"Namespace":"app-ns","Name":"policy-2"}]`},
-			},
-		},
-		PolicyRefsConfig: &KuadrantRateLimitPolicyRefsConfig{},
-	}
-	if gw.PolicyRefsAnnotation() != RateLimitPoliciesBackRefAnnotation {
-		t.Fail()
-	}
-}
-
 func TestGetGatewayWorkloadSelector(t *testing.T) {
 	hostnameAddress := gatewayapiv1beta1.AddressType("Hostname")
 	gateway := &gatewayapiv1beta1.Gateway{
@@ -1197,175 +817,4 @@ func TestValidateHierarchicalRules(t *testing.T) {
 		t.Fatal("the error message does not match the expected error one", expectedError.Error(), err.Error())
 	}
 
-}
-
-func TestIsHTTPRouteAccepted(t *testing.T) {
-	testCases := []struct {
-		name     string
-		route    *gatewayapiv1beta1.HTTPRoute
-		expected bool
-	}{
-		{
-			"nil",
-			nil,
-			false,
-		},
-		{
-			"empty parent refs",
-			&gatewayapiv1beta1.HTTPRoute{
-				Spec: gatewayapiv1beta1.HTTPRouteSpec{},
-			},
-			false,
-		},
-		{
-			"single parent accepted",
-			&gatewayapiv1beta1.HTTPRoute{
-				Spec: gatewayapiv1beta1.HTTPRouteSpec{
-					CommonRouteSpec: gatewayapiv1beta1.CommonRouteSpec{
-						ParentRefs: []gatewayapiv1beta1.ParentReference{
-							{
-								Name: "a",
-							},
-						},
-					},
-				},
-				Status: gatewayapiv1beta1.HTTPRouteStatus{
-					RouteStatus: gatewayapiv1beta1.RouteStatus{
-						Parents: []gatewayapiv1beta1.RouteParentStatus{
-							{
-								ParentRef: gatewayapiv1beta1.ParentReference{
-									Name: "a",
-								},
-								Conditions: []metav1.Condition{
-									{
-										Type:   "Accepted",
-										Status: metav1.ConditionTrue,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			true,
-		},
-		{
-			"single parent not accepted",
-			&gatewayapiv1beta1.HTTPRoute{
-				Spec: gatewayapiv1beta1.HTTPRouteSpec{
-					CommonRouteSpec: gatewayapiv1beta1.CommonRouteSpec{
-						ParentRefs: []gatewayapiv1beta1.ParentReference{
-							{
-								Name: "a",
-							},
-						},
-					},
-				},
-				Status: gatewayapiv1beta1.HTTPRouteStatus{
-					RouteStatus: gatewayapiv1beta1.RouteStatus{
-						Parents: []gatewayapiv1beta1.RouteParentStatus{
-							{
-								ParentRef: gatewayapiv1beta1.ParentReference{
-									Name: "a",
-								},
-								Conditions: []metav1.Condition{
-									{
-										Type:   "Accepted",
-										Status: metav1.ConditionFalse,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			false,
-		},
-		{
-			"wrong parent is accepted",
-			&gatewayapiv1beta1.HTTPRoute{
-				Spec: gatewayapiv1beta1.HTTPRouteSpec{
-					CommonRouteSpec: gatewayapiv1beta1.CommonRouteSpec{
-						ParentRefs: []gatewayapiv1beta1.ParentReference{
-							{
-								Name: "a",
-							},
-						},
-					},
-				},
-				Status: gatewayapiv1beta1.HTTPRouteStatus{
-					RouteStatus: gatewayapiv1beta1.RouteStatus{
-						Parents: []gatewayapiv1beta1.RouteParentStatus{
-							{
-								ParentRef: gatewayapiv1beta1.ParentReference{
-									Name: "b",
-								},
-								Conditions: []metav1.Condition{
-									{
-										Type:   "Accepted",
-										Status: metav1.ConditionTrue,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			false,
-		},
-		{
-			"multiple parents only one is accepted",
-			&gatewayapiv1beta1.HTTPRoute{
-				Spec: gatewayapiv1beta1.HTTPRouteSpec{
-					CommonRouteSpec: gatewayapiv1beta1.CommonRouteSpec{
-						ParentRefs: []gatewayapiv1beta1.ParentReference{
-							{
-								Name: "a",
-							},
-							{
-								Name: "b",
-							},
-						},
-					},
-				},
-				Status: gatewayapiv1beta1.HTTPRouteStatus{
-					RouteStatus: gatewayapiv1beta1.RouteStatus{
-						Parents: []gatewayapiv1beta1.RouteParentStatus{
-							{
-								ParentRef: gatewayapiv1beta1.ParentReference{
-									Name: "a",
-								},
-								Conditions: []metav1.Condition{
-									{
-										Type:   "Accepted",
-										Status: metav1.ConditionTrue,
-									},
-								},
-							},
-							{
-								ParentRef: gatewayapiv1beta1.ParentReference{
-									Name: "b",
-								},
-								Conditions: []metav1.Condition{
-									{
-										Type:   "Accepted",
-										Status: metav1.ConditionFalse,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			false,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(subT *testing.T) {
-			res := IsHTTPRouteAccepted(tc.route)
-			if res != tc.expected {
-				subT.Errorf("result (%t) does not match expected (%t)", res, tc.expected)
-			}
-		})
-	}
 }
