+++ pkg/common/referrer_test.go
@@ -1,3 +1,5 @@
+// TODO: move to https://github.com/Kuadrant/gateway-api-machinery
+
 package common
 
 import (
