+++ pkg/mappers/event_mapper.go
@@ -1,3 +1,4 @@
+// TODO: move to https://github.com/Kuadrant/gateway-api-machinery
 package mappers
 
 import (
