+++ pkg/mappers/gateway_test.go
@@ -1,9 +1,10 @@
+// TODO: move to https://github.com/Kuadrant/gateway-api-machinery
 package mappers
 
 import (
 	"testing"
 )
 
-func TestNewGatewayEventMapper(t *testing.T) {
+func TestNewGatewayEventMapper(_ *testing.T) {
 	_ = NewGatewayEventMapper()
 }
