+++ pkg/mappers/httproute_test.go
@@ -1,9 +1,10 @@
+// TODO: move to https://github.com/Kuadrant/gateway-api-machinery
 package mappers
 
 import (
 	"testing"
 )
 
-func TestNewHTTPRouteEventMapper(t *testing.T) {
+func TestNewHTTPRouteEventMapper(_ *testing.T) {
 	_ = NewHTTPRouteEventMapper()
 }
