+++ pkg/reconcilers/targetref_reconciler.go
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+// TODO: move to https://github.com/Kuadrant/gateway-api-machinery
 package reconcilers
 
 import (
@@ -26,7 +27,6 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	gatewayapiv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 	gatewayapiv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
 	"github.com/kuadrant/kuadrant-operator/pkg/common"
@@ -41,58 +41,6 @@ var _ reconcile.Reconciler = &TargetRefReconciler{}
 
 func (r *TargetRefReconciler) Reconcile(context.Context, ctrl.Request) (ctrl.Result, error) {
 	return reconcile.Result{}, nil
-}
-
-func (r *TargetRefReconciler) FetchValidGateway(ctx context.Context, key client.ObjectKey) (*gatewayapiv1beta1.Gateway, error) {
-	logger, _ := logr.FromContext(ctx)
-
-	gw := &gatewayapiv1beta1.Gateway{}
-	err := r.Client().Get(ctx, key, gw)
-	logger.V(1).Info("FetchValidGateway", "gateway", key, "err", err)
-	if err != nil {
-		return nil, err
-	}
-
-	if meta.IsStatusConditionFalse(gw.Status.Conditions, common.GatewayProgrammedConditionType) {
-		return nil, fmt.Errorf("FetchValidGateway: gateway (%v) not ready", key)
-	}
-
-	return gw, nil
-}
-
-func (r *TargetRefReconciler) FetchValidHTTPRoute(ctx context.Context, key client.ObjectKey) (*gatewayapiv1beta1.HTTPRoute, error) {
-	logger, _ := logr.FromContext(ctx)
-
-	httpRoute := &gatewayapiv1beta1.HTTPRoute{}
-	err := r.Client().Get(ctx, key, httpRoute)
-	logger.V(1).Info("FetchValidHTTPRoute", "httpRoute", key, "err", err)
-	if err != nil {
-		return nil, err
-	}
-
-	if !common.IsHTTPRouteAccepted(httpRoute) {
-		return nil, fmt.Errorf("FetchValidHTTPRoute: httproute (%v) not accepted", key)
-	}
-
-	return httpRoute, nil
-}
-
-// FetchValidTargetRef fetches the target reference object and checks the status is valid
-func (r *TargetRefReconciler) FetchValidTargetRef(ctx context.Context, targetRef gatewayapiv1alpha2.PolicyTargetReference, defaultNs string) (client.Object, error) {
-	tmpNS := defaultNs
-	if targetRef.Namespace != nil {
-		tmpNS = string(*targetRef.Namespace)
-	}
-
-	objKey := client.ObjectKey{Name: string(targetRef.Name), Namespace: tmpNS}
-
-	if common.IsTargetRefHTTPRoute(targetRef) {
-		return r.FetchValidHTTPRoute(ctx, objKey)
-	} else if common.IsTargetRefGateway(targetRef) {
-		return r.FetchValidGateway(ctx, objKey)
-	}
-
-	return nil, fmt.Errorf("FetchValidTargetRef: targetRef (%v) to unknown network resource", targetRef)
 }
 
 // FetchAcceptedGatewayHTTPRoutes returns the list of HTTPRoutes that have been accepted as children of a gateway.
@@ -203,7 +151,7 @@ func (r *TargetRefReconciler) DeleteTargetBackReference(ctx context.Context, tar
 // this list of policy refs; nevertheless, the actual order of returned policy refs depends on the order the policy refs
 // appear in the annotations of the gateways.
 // Only gateways with status programmed are considered.
-func (r *TargetRefReconciler) GetAllGatewayPolicyRefs(ctx context.Context, policyRefsConfig common.PolicyRefsConfig) ([]client.ObjectKey, error) {
+func (r *TargetRefReconciler) GetAllGatewayPolicyRefs(ctx context.Context, policyRefsConfig common.Referrer) ([]client.ObjectKey, error) {
 	var uniquePolicyRefs map[string]struct{}
 	var policyRefs []client.ObjectKey
 
@@ -213,19 +161,19 @@ func (r *TargetRefReconciler) GetAllGatewayPolicyRefs(ctx context.Context, polic
 	}
 
 	// sort the gateways by creation timestamp to mitigate the risk of non-idenpotent reconciliations
-	var gateways common.GatewayWrapperList
+	var gateways GatewayWrapperList
 	for i := range gwList.Items {
 		gateway := gwList.Items[i]
 		// skip gateways that are not managed by kuadrant or that are not ready
 		if !common.IsKuadrantManaged(&gateway) || meta.IsStatusConditionFalse(gateway.Status.Conditions, common.GatewayProgrammedConditionType) {
 			continue
 		}
-		gateways = append(gateways, common.GatewayWrapper{Gateway: &gateway, PolicyRefsConfig: policyRefsConfig})
+		gateways = append(gateways, GatewayWrapper{Gateway: &gateway, Referrer: policyRefsConfig})
 	}
 	sort.Sort(gateways)
 
 	for _, gw := range gateways {
-		for _, policyRef := range gw.PolicyRefs() {
+		for _, policyRef := range common.BackReferencesFromObject(gw.Gateway, gw.Referrer) {
 			if _, ok := uniquePolicyRefs[policyRef.String()]; ok {
 				continue
 			}
@@ -236,43 +184,9 @@ func (r *TargetRefReconciler) GetAllGatewayPolicyRefs(ctx context.Context, polic
 	return policyRefs, nil
 }
 
-// Returns:
-// * list of gateways to which the policy applies for the first time
-// * list of gateways to which the policy no longer applies
-// * list of gateways to which the policy still applies
-func (r *TargetRefReconciler) ComputeGatewayDiffs(ctx context.Context, policy common.KuadrantPolicy, targetNetworkObject client.Object, policyRefsConfig common.PolicyRefsConfig) (*GatewayDiff, error) {
-	logger, _ := logr.FromContext(ctx)
-
-	var gwKeys []client.ObjectKey
-	if policy.GetDeletionTimestamp() == nil {
-		gwKeys = r.TargetedGatewayKeys(ctx, targetNetworkObject)
-	}
-
-	// TODO(rahulanand16nov): maybe think about optimizing it with a label later
-	allGwList := &gatewayapiv1beta1.GatewayList{}
-	err := r.Client().List(ctx, allGwList)
-	if err != nil {
-		return nil, err
-	}
-
-	gwDiff := &GatewayDiff{
-		GatewaysMissingPolicyRef:     common.GatewaysMissingPolicyRef(allGwList, client.ObjectKeyFromObject(policy), gwKeys, policyRefsConfig),
-		GatewaysWithValidPolicyRef:   common.GatewaysWithValidPolicyRef(allGwList, client.ObjectKeyFromObject(policy), gwKeys, policyRefsConfig),
-		GatewaysWithInvalidPolicyRef: common.GatewaysWithInvalidPolicyRef(allGwList, client.ObjectKeyFromObject(policy), gwKeys, policyRefsConfig),
-	}
-
-	logger.V(1).Info("ComputeGatewayDiffs",
-		"#missing-policy-ref", len(gwDiff.GatewaysMissingPolicyRef),
-		"#valid-policy-ref", len(gwDiff.GatewaysWithValidPolicyRef),
-		"#invalid-policy-ref", len(gwDiff.GatewaysWithInvalidPolicyRef),
-	)
-
-	return gwDiff, nil
-}
-
 // ReconcileGatewayPolicyReferences updates the annotations in the Gateway resources that list to all the policies
 // that directly or indirectly target the gateway, based upon a pre-computed gateway diff object
-func (r *TargetRefReconciler) ReconcileGatewayPolicyReferences(ctx context.Context, policy client.Object, gwDiffObj *GatewayDiff) error {
+func (r *TargetRefReconciler) ReconcileGatewayPolicyReferences(ctx context.Context, policy client.Object, gwDiffObj *GatewayDiffs) error {
 	logger, _ := logr.FromContext(ctx)
 
 	// delete the policy from the annotations of the gateways no longer target by the policy
@@ -298,10 +212,4 @@ func (r *TargetRefReconciler) ReconcileGatewayPolicyReferences(ctx context.Conte
 	}
 
 	return nil
-}
-
-type GatewayDiff struct {
-	GatewaysMissingPolicyRef     []common.GatewayWrapper
-	GatewaysWithValidPolicyRef   []common.GatewayWrapper
-	GatewaysWithInvalidPolicyRef []common.GatewayWrapper
 }
