+++ pkg/reconcilers/targetref_reconciler_test.go
@@ -16,11 +16,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+// TODO: move to https://github.com/Kuadrant/gateway-api-machinery
 package reconcilers
 
 import (
 	"context"
-	"reflect"
 	"testing"
 
 	"github.com/go-logr/logr"
@@ -31,257 +31,10 @@ import (
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	gatewayapiv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 	gatewayapiv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
 	"github.com/kuadrant/kuadrant-operator/pkg/log"
 )
-
-func TestFetchValidGateway(t *testing.T) {
-	var (
-		namespace = "operator-unittest"
-		gwName    = "my-gateway"
-	)
-	baseCtx := context.Background()
-	ctx := logr.NewContext(baseCtx, log.Log)
-
-	s := scheme.Scheme
-	err := appsv1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = gatewayapiv1beta1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	existingGateway := &gatewayapiv1beta1.Gateway{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "gateway.networking.k8s.io/v1beta1",
-			Kind:       "Gateway",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      gwName,
-			Namespace: namespace,
-		},
-		Spec: gatewayapiv1beta1.GatewaySpec{
-			GatewayClassName: "istio",
-		},
-		Status: gatewayapiv1beta1.GatewayStatus{
-			Conditions: []metav1.Condition{
-				{
-					Type:   "Ready",
-					Status: metav1.ConditionTrue,
-				},
-			},
-		},
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existingGateway}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	clientAPIReader := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	recorder := record.NewFakeRecorder(1000)
-
-	baseReconciler := NewBaseReconciler(cl, s, clientAPIReader, log.Log, recorder)
-	targetRefReconciler := TargetRefReconciler{
-		BaseReconciler: baseReconciler,
-	}
-
-	key := client.ObjectKey{Name: gwName, Namespace: namespace}
-
-	res, err := targetRefReconciler.FetchValidGateway(ctx, key)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if res == nil {
-		t.Fatal("res is nil")
-	}
-
-	if !reflect.DeepEqual(res.Spec, existingGateway.Spec) {
-		t.Fatal("res spec not as expected")
-	}
-}
-
-func TestFetchValidHTTPRoute(t *testing.T) {
-	var (
-		namespace = "operator-unittest"
-		routeName = "my-route"
-	)
-	baseCtx := context.Background()
-	ctx := logr.NewContext(baseCtx, log.Log)
-
-	s := scheme.Scheme
-	err := appsv1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = gatewayapiv1beta1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	existingRoute := &gatewayapiv1beta1.HTTPRoute{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "gateway.networking.k8s.io/v1beta1",
-			Kind:       "HTTPRoute",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      routeName,
-			Namespace: namespace,
-		},
-		Spec: gatewayapiv1beta1.HTTPRouteSpec{
-			CommonRouteSpec: gatewayapiv1beta1.CommonRouteSpec{
-				ParentRefs: []gatewayapiv1beta1.ParentReference{
-					{
-						Name: "gwName",
-					},
-				},
-			},
-		},
-		Status: gatewayapiv1beta1.HTTPRouteStatus{
-			RouteStatus: gatewayapiv1beta1.RouteStatus{
-				Parents: []gatewayapiv1beta1.RouteParentStatus{
-					{
-						ParentRef: gatewayapiv1beta1.ParentReference{
-							Name: "gwName",
-						},
-						Conditions: []metav1.Condition{
-							{
-								Type:   "Accepted",
-								Status: metav1.ConditionTrue,
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existingRoute}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	clientAPIReader := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	recorder := record.NewFakeRecorder(1000)
-
-	baseReconciler := NewBaseReconciler(cl, s, clientAPIReader, log.Log, recorder)
-	targetRefReconciler := TargetRefReconciler{
-		BaseReconciler: baseReconciler,
-	}
-
-	key := client.ObjectKey{Name: routeName, Namespace: namespace}
-
-	res, err := targetRefReconciler.FetchValidHTTPRoute(ctx, key)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if res == nil {
-		t.Fatal("res is nil")
-	}
-
-	if !reflect.DeepEqual(res.Spec, existingRoute.Spec) {
-		t.Fatal("res spec not as expected")
-	}
-}
-
-func TestFetchValidTargetRef(t *testing.T) {
-	var (
-		namespace = "operator-unittest"
-		routeName = "my-route"
-	)
-	baseCtx := context.Background()
-	ctx := logr.NewContext(baseCtx, log.Log)
-
-	s := scheme.Scheme
-	err := appsv1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = gatewayapiv1beta1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	targetRef := gatewayapiv1alpha2.PolicyTargetReference{
-		Group: "gateway.networking.k8s.io",
-		Kind:  "HTTPRoute",
-		Name:  gatewayapiv1beta1.ObjectName(routeName),
-	}
-
-	existingRoute := &gatewayapiv1beta1.HTTPRoute{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "gateway.networking.k8s.io/v1beta1",
-			Kind:       "HTTPRoute",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      routeName,
-			Namespace: namespace,
-		},
-		Spec: gatewayapiv1beta1.HTTPRouteSpec{
-			CommonRouteSpec: gatewayapiv1beta1.CommonRouteSpec{
-				ParentRefs: []gatewayapiv1beta1.ParentReference{
-					{
-						Name: "gwName",
-					},
-				},
-			},
-		},
-		Status: gatewayapiv1beta1.HTTPRouteStatus{
-			RouteStatus: gatewayapiv1beta1.RouteStatus{
-				Parents: []gatewayapiv1beta1.RouteParentStatus{
-					{
-						ParentRef: gatewayapiv1beta1.ParentReference{
-							Name: "gwName",
-						},
-						Conditions: []metav1.Condition{
-							{
-								Type:   "Accepted",
-								Status: metav1.ConditionTrue,
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existingRoute}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	clientAPIReader := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	recorder := record.NewFakeRecorder(1000)
-
-	baseReconciler := NewBaseReconciler(cl, s, clientAPIReader, log.Log, recorder)
-	targetRefReconciler := TargetRefReconciler{
-		BaseReconciler: baseReconciler,
-	}
-
-	res, err := targetRefReconciler.FetchValidTargetRef(ctx, targetRef, namespace)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if res == nil {
-		t.Fatal("res is nil")
-	}
-
-	switch obj := res.(type) {
-	case *gatewayapiv1beta1.HTTPRoute:
-		if !reflect.DeepEqual(obj.Spec, existingRoute.Spec) {
-			t.Fatal("res spec not as expected")
-		}
-	default:
-		t.Fatal("res type not known")
-	}
-}
 
 func TestReconcileTargetBackReference(t *testing.T) {
 	var (
@@ -380,87 +133,6 @@ func TestReconcileTargetBackReference(t *testing.T) {
 
 	if val != policyKey.String() {
 		t.Fatalf("annotation value (%s) does not match expected (%s)", val, policyKey.String())
-	}
-}
-
-func TestTargetedGatewayKeys(t *testing.T) {
-	var (
-		namespace = "operator-unittest"
-		routeName = "my-route"
-	)
-	baseCtx := context.Background()
-	ctx := logr.NewContext(baseCtx, log.Log)
-
-	s := scheme.Scheme
-	err := appsv1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = gatewayapiv1beta1.AddToScheme(s)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	existingRoute := &gatewayapiv1beta1.HTTPRoute{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "gateway.networking.k8s.io/v1beta1",
-			Kind:       "HTTPRoute",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      routeName,
-			Namespace: namespace,
-		},
-		Spec: gatewayapiv1beta1.HTTPRouteSpec{
-			CommonRouteSpec: gatewayapiv1beta1.CommonRouteSpec{
-				ParentRefs: []gatewayapiv1beta1.ParentReference{
-					{
-						Name: "gwName",
-					},
-				},
-			},
-		},
-		Status: gatewayapiv1beta1.HTTPRouteStatus{
-			RouteStatus: gatewayapiv1beta1.RouteStatus{
-				Parents: []gatewayapiv1beta1.RouteParentStatus{
-					{
-						ParentRef: gatewayapiv1beta1.ParentReference{
-							Name: "gwName",
-						},
-						Conditions: []metav1.Condition{
-							{
-								Type:   "Accepted",
-								Status: metav1.ConditionTrue,
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Objects to track in the fake client.
-	objs := []runtime.Object{existingRoute}
-
-	// Create a fake client to mock API calls.
-	cl := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	clientAPIReader := fake.NewClientBuilder().WithRuntimeObjects(objs...).Build()
-	recorder := record.NewFakeRecorder(1000)
-
-	baseReconciler := NewBaseReconciler(cl, s, clientAPIReader, log.Log, recorder)
-	targetRefReconciler := TargetRefReconciler{
-		BaseReconciler: baseReconciler,
-	}
-
-	keys := targetRefReconciler.TargetedGatewayKeys(ctx, existingRoute)
-
-	if len(keys) != 1 {
-		t.Fatalf("gateway key slice length is %d and it was expected to be 1", len(keys))
-	}
-
-	expectedKey := client.ObjectKey{Name: "gwName", Namespace: namespace}
-
-	if keys[0] != expectedKey {
-		t.Fatalf("gwKey value (%+v) does not match expected (%+v)", keys[0], expectedKey)
 	}
 }
